+++ wio/ThermometerAndHumidity.cpp
@@ -9,24 +9,24 @@ const char* ThermometerAndHumidity::ConvertReadings()
     // problems of loosing the data once the 'readingsBuffer' goes out of scope.
     char* readingsBuffer = new char[50];
 
-    long temperatures = GetTemperature();
-    long humidity = GetHumidity();
+    float temperatures = GetTemperature();
+    float humidity = GetHumidity();
 
     std::sprintf(readingsBuffer, "%ldC %ld%%", temperatures, humidity);
     
     return readingsBuffer;
 }
 
-long ThermometerAndHumidity::GetTemperature()
+float ThermometerAndHumidity::GetTemperature()
 {
-    long temperature = _dht.readTemperature();
+    float temperature = _dht.readTemperature();
 
     return temperature;
 }
 
-long ThermometerAndHumidity::GetHumidity()
+float ThermometerAndHumidity::GetHumidity()
 {
-    long humidity = _dht.readHumidity();
+    float humidity = _dht.readHumidity();
 
     return humidity;
 }
